Move the CLI helper code from a hardwired string into a separate .py file
